requirements.txt
tensorflow>=2.19.0
keras>=2.8.0
flask>=2.0.0
numpy>=1.19.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pillow>=8.0.0
pandas>=1.3.0
werkzeug>=2.0.0

